fix: mouse cursor remain in auto-scrolling mode
When user clicks the card with the middle mouse button,
mouseDown event is fired adnd the cursor transistions to
auto-scrolling mode and a new page opens in new tab.
No matter where you click on the newly opened tab,
it will not transistion back to normal mouse cursor.

This code change prevents the auto-scrolling to engage
by calling the event.preventDefault function.