fix: TwohopLink header opens the file with the mouse down event.
After a file is opened from the TwohopLink header with the mouse down
event, if there is a card under the cursor when user releases the
button, the page of the card that happens to be under the cursor will
open. Resulting in an unexpected hop.

This code change prevents this behavior by removing the onMouseDown from
the twohop-links-twohop-header